requirements.txt
# Sprint Report CLI - Python Dependencies

# Core
anthropic==0.39.0
requests==2.31.0
python-dotenv==1.0.0
PyYAML==6.0.1

# CLI Interface
rich==13.7.0

# PDF Generation
weasyprint==62.3
Jinja2==3.1.3
markdown==3.5.2

# Utilities
python-dateutil==2.8.2

# Data Validation
pydantic==2.5.3
